feature(editorialcontrol-site): Phase 1 - multi-site source layout and build split
Split the single Astro build into two per-site builds so editorialcontrol.org
can be added as a sibling to audiocontrol.org from the same repo.

Source layout:
- Move src/pages, src/layouts, src/components, src/styles under
  src/sites/audiocontrol/. Moving the shared dirs too (not just pages) keeps
  all existing relative imports valid without rewriting ~20 pages; Phase 3's
  brand refactor can extract shared base layouts back to src/layouts/ when
  editorialcontrol actually needs them.
- Add src/sites/editorialcontrol/pages/index.astro as a placeholder.

Configs:
- astro.config.mjs -> astro.audiocontrol.config.mjs + astro.editorialcontrol.config.mjs,
  each with its own site, srcDir, outDir, and sitemap config.
- audiocontrol outputs to dist/audiocontrol; editorialcontrol to dist/editorialcontrol.

Scripts and deploy:
- package.json: dev, dev:audiocontrol, dev:editorialcontrol, build (runs both),
  build:audiocontrol, build:editorialcontrol, preview and preview:* variants.
- netlify.toml: audiocontrol Netlify site now runs build:audiocontrol and
  publishes dist/audiocontrol. Editorialcontrol's Netlify site will be created
  in the UI at Phase 6 / Launch.

Path fixes (required by the depth change):
- src/sites/audiocontrol/pages/api/dev/feature-image/{generate,log,workflow}.ts
  import paths and __dirname math updated from 5 to 7 levels to reach repo root.
- src/sites/audiocontrol/pages/dev/feature-image-preview.astro filter import
  extended by two levels.

Verification:
- Captured a pre-split dist baseline, rebuilt, and ran a normalized diff.
  All HTML, sitemap, and asset outputs are byte-identical after normalizing
  Astro's auto-generated data-astro-cid-* hashes (which are derived from
  source paths and changed when files moved). CSS content pairs all match
  their pre-split counterparts.
- editorialcontrol build produces a working placeholder page at
  dist/editorialcontrol/index.html with favicons, fonts, and sitemap.
- npm test: 93 pass, 2 pre-existing failures (integration tests that hit the
  live Roland editor Netlify app; unrelated to this change).

Refs: GitHub #69 (parent), #70 (Phase 1).
