+++ docs/1.0/001-IN-PROGRESS/editorialcontrol-site/README.md
@@ -13,7 +13,7 @@ Stand up editorialcontrol.org as a sibling site to audiocontrol.org. One repo, t
 
 | Phase | Description | Status |
 |-------|-------------|--------|
-| 1 | Multi-site source layout + build split | Not Started |
+| 1 | Multi-site source layout + build split | Implementation complete (Netlify UI work deferred to Phase 6) |
 | 2 | Multi-site editorial calendar library | Not Started |
 | 3 | editorialcontrol branding + core pages | Not Started |
 | 4 | Content migration (2 posts) | Not Started |
