+++ netlify.toml
@@ -1,6 +1,6 @@
 [build]
-  command = "npm run build"
-  publish = "dist"
+  command = "npm run build:audiocontrol"
+  publish = "dist/audiocontrol"
 
 # Sitemap alias
 [[redirects]]
