+++ package.json
@@ -3,9 +3,15 @@
   "type": "module",
   "version": "0.0.1",
   "scripts": {
-    "dev": "astro dev",
-    "build": "npm run generate-og && astro build",
-    "preview": "astro preview",
+    "dev": "astro dev --config astro.audiocontrol.config.mjs",
+    "dev:audiocontrol": "astro dev --config astro.audiocontrol.config.mjs",
+    "dev:editorialcontrol": "astro dev --config astro.editorialcontrol.config.mjs",
+    "build": "npm run build:audiocontrol && npm run build:editorialcontrol",
+    "build:audiocontrol": "npm run generate-og && astro build --config astro.audiocontrol.config.mjs",
+    "build:editorialcontrol": "astro build --config astro.editorialcontrol.config.mjs",
+    "preview": "astro preview --config astro.audiocontrol.config.mjs",
+    "preview:audiocontrol": "astro preview --config astro.audiocontrol.config.mjs",
+    "preview:editorialcontrol": "astro preview --config astro.editorialcontrol.config.mjs",
     "astro": "astro",
     "generate-og": "tsx scripts/generate-og-images.ts",
     "generate-feature-image": "tsx scripts/feature-image/cli.ts",
