+++ src/sites/audiocontrol/components/SpecsTable.astro
