+++ src/sites/audiocontrol/pages/blog/reverse-engineering-akai-s3000xl-midi-over-scsi/index.md
